preprocessing/py27_requirements.txt
# After these ones, install py27_requirements_dependent_on_others.txt
acme==0.4.1
attrs==15.2.0
beautifulsoup4==4.4.1
biopython==1.74
blinker==1.3
chardet==2.3.0
ConfigArgParse==0.10.0
configobj==5.0.6
cryptography==3.3.2
decorator==4.0.6
dnspython==1.12.0
enum34==1.1.2
ete3==3.0.0b35
fcsparser==0.2.0
FlowCytometryTools==0.5.0
funcsigs==1.0.2
future==0.15.2
gmpy==1.17
html5lib==0.999
httplib2==0.9.1
idna==2.0
ipaddress==1.0.16
ipython==3.1.0
Jinja2==2.8
joblib==0.14.1
letsencrypt==0.4.1
letsencrypt-apache==0.4.1
llvmlite==0.23.0
lockfile==0.12.2
lxml==3.5.0
MarkupSafe==0.23
matplotlib==1.5.3
mercurial==3.7.3
mock==1.3.0
mpmath==0.19
ndg-httpsclient==0.4.0
networkx==1.11
nose==1.3.7
numba==0.38.0
numexpr==2.6.8
numpy==1.15.4
oauthlib==1.0.3
openpyxl==2.3.0
pandas==0.24.2
parsedatetime==1.4
patsy==0.4.1
pbr==1.8.0
pexpect==4.0.1
Pillow==3.1.2
piston-mini-client==0.7.5
psutil==3.4.2
# Psycopg2: see https://github.com/psycopg/psycopg2-wheels/issues/2 and  https://github.com/psycopg/psycopg2/issues/543
psycopg2==2.7.7 --no-binary psycopg2
ptyprocess==0.5
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pydotplus==2.0.2
pyglet==1.1.4
Pygments==2.0.2
pygraphviz==1.3.1
PyJWT==1.3.0
pyliblzma==0.5.3
PyMySQL==0.7.2
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyRFC3339==1.0
pysam==0.9.0
pyserial==3.0.1
pytest==2.8.7
python-augeas==0.5.0
python-dateutil==2.7.3
python2-pythondialog==3.3.0
pytz==2018.5
PyVCF==0.6.7
PyX==0.12.1
pyxdg==0.25
PyYAML==3.11
pyzmq==15.2.0
reportlab==3.3.0
requests==2.9.1
scikit-learn==0.20.0
scipy==1.1.0
scour==0.32
seaborn==0.6.0
service-identity==16.0.0
simplegeneric==0.8.1
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.1.10
statsmodels==0.6.1
sympy==0.7.6.1
tables==3.2.2
tornado==4.2.1
Twisted==16.0.0
umap-learn==0.2.3
urllib3==1.13.1
xlrd==0.9.4
xlwt==0.7.5
zmq==0.0.0
zope.component==4.2.2
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.1.3
